Work on HUDFX + Folder Cleanup
Removed unused scripts and scenes to avoid confusion.
Worked on hooking up the HUD effects to the existing health system. Not fully operational yet, needs to be further worked on
