--- Assets/_Developers/AP/PaulS/ConfirmationPopUps.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e5e73fe9f40f7f44191f7c515cb4f935
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
