--- Assets/_Developers/AP/PaulS/ConfirmationWindow.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-using TMPro;
-
-public class ConfirmationWindow : MonoBehaviour
-{
-    public Button yesButton;
-    public Button noButton;
-    public TextMeshProUGUI messageText;
-}
