--- Assets/_Developers/AP/PaulS/ConfirmationWindow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0d6d391aeedcdfa4fb5b18ac59591262
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
