--- Assets/_Developers/AP/PaulS/HealthController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd92da77d2bb0904cb51ada6730e0a18
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
