+++ Assets/_Developers/AP/PaulS/HealthHUD.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 76b2da4d6580de44886edc962d984079
-DefaultImporter:
+guid: b4270fdabf0f25447b11f81672abfa4e
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
