--- Assets/_Developers/AP/PaulS/TestScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4e0978ff349d2fc44a6373098c2327a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
