--- Assets/_Developers/AP/PaulS/objectiveTimer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0c2d8444fdf073149a9d176128e96bc0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
